delete timezones
