+++ common/admin.py
@@ -5,7 +5,7 @@ from .models import City
 
 
 class CityAdmin(StaffRequiredAdminMixin, admin.ModelAdmin):
-    list_display = ("id", "name", "isPrimary", "timeZone")
+    list_display = ("id", "name", "isPrimary")
     list_display_links = ("name",)
     search_fields = ("name",)
     list_filter = ("isPrimary",)
