+++ common/views.py
@@ -4,8 +4,8 @@ from rest_framework.mixins import ListModelMixin, UpdateModelMixin
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.viewsets import GenericViewSet
 
-from common.models import City, Tag
-from common.serializers import CitySerializer, MyCitySerializer, TagSerializer
+from common.models import City
+from common.serializers import CitySerializer, MyCitySerializer
 from users.models import Profile
 
 
@@ -23,8 +23,3 @@ class MyCityApiView(ListModelMixin, UpdateModelMixin, GenericViewSet):
         user = self.request.user
         queryset = Profile.objects.filter(user=user)
         return queryset
-
-
-class TagList(ListAPIView):
-    queryset = Tag.objects.all().order_by("-name")
-    serializer_class = TagSerializer
