+++ config/settings/base.py
@@ -53,7 +53,6 @@ MIDDLEWARE = [
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
-    'common.middleware.set_timezone',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
