+++ questions/management/slug_transliteration.py
@@ -1,0 +1,39 @@
+from django.template.defaultfilters import slugify as django_slugify
+
+ALPHABET = {
+    "а": "a",
+    "б": "b",
+    "в": "v",
+    "г": "g",
+    "д": "d",
+    "е": "e",
+    "ё": "yo",
+    "ж": "zh",
+    "з": "z",
+    "и": "i",
+    "й": "j",
+    "к": "k",
+    "л": "l",
+    "м": "m",
+    "н": "n",
+    "о": "o",
+    "п": "p",
+    "р": "r",
+    "с": "s",
+    "т": "t",
+    "у": "u",
+    "ф": "f",
+    "х": "kh",
+    "ц": "ts",
+    "ч": "ch",
+    "ш": "sh",
+    "щ": "shch",
+    "ы": "i",
+    "э": "e",
+    "ю": "yu",
+    "я": "ya",
+}
+
+
+def slugify(name):
+    return django_slugify("".join(ALPHABET.get(w, w) for w in name.lower()))
