+++ questions/models.py
@@ -1,12 +1,22 @@
-from common.models import Tag
 from django.db import models
+
+
+class Tag(models.Model):
+    name = models.CharField(max_length=50, unique=True)
+    slug = models.SlugField(max_length=50, unique=True)
+
+    class Meta:
+        verbose_name = "Тег"
+        verbose_name_plural = "Теги"
+
+    def __str__(self):
+        return self.name
 
 
 class Question(models.Model):
     tag = models.ManyToManyField(
         Tag,
         related_name="tags",
-        null=True,
         blank=True,
     )
     question = models.CharField(max_length=500, unique=True)
